pyproject.toml
[tool.poetry]
name = "quiz"
version = "0.1.0"
description = ""
authors = ["youngwishes <d.fedorov@ylab.team>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.0"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"
django = "^5.0.3"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
